pyproject.toml
[tool.poetry]
name = "db-write-performance-benchmark"
version = "0.1.0"
description = ""
authors = ["Karn Wong <karn@karnwong.me>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11,<3.13"
pyspark = "^3.5.1"
pandas = "^2.2.1"
polars = "^0.20.15"
dlt = "^0.4.6"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"
psutil = "^5.9.8"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"


[tool.poetry.group.tests.dependencies]
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
